models/VILA_whisper/AutoGaze/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autogaze"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "torch",
    "torchvision",
    "flash_attn",
    "hydra-core>=1.3.2",
    "wandb",
    "loguru",
    "timm>=1.0.15",
    "tqdm",
    "transformers~=4.51",
    "pillow",
    "numpy",
    "omegaconf",
    "matplotlib",
    "einops",
    "av",
    "imageio"   
]

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
